+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Allow single valued `dimIndex`
 - Added `reg_iter`, `reg_iter_mut` methods on `Peripheral` and `Cluster`
 - Added `DerefMut` for `Cluster`, `Register` and `Field`
 - Added `display_name` to `RegisterInfo`